refactor relaxometry task
